refactor: move deployment options under types
